requirements.txt
streamlit==1.51.0
ultralytics==8.3.228
pandas==2.3.3
numpy==2.2.6
pillow==12.0.0
plotly==6.5.0
easyocr
opencv-python-headless==4.11.0.86
scikit-learn==1.8.0
mplfinance
sahi==0.11.36
python-dateutil==2.9.0.post0
python-bidi==0.6.7
